refactor: 파일명 및 파일 구조 변경 #25
