+++ FE/app/src/main/java/com/wonchihyeon/ytt_android/MainActivity.kt
@@ -3,10 +3,10 @@ package com.wonchihyeon.ytt_android
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.wonchihyeon.ytt_android.databinding.ActivityMainBinding
-import com.wonchihyeon.ytt_android.fragments.FavoriteFragment
-import com.wonchihyeon.ytt_android.fragments.HomeFragment
-import com.wonchihyeon.ytt_android.fragments.MyFragment
-import com.wonchihyeon.ytt_android.fragments.OrderListFragment
+import com.wonchihyeon.ytt_android.view.FavoriteFragment
+import com.wonchihyeon.ytt_android.view.HomeFragment
+import com.wonchihyeon.ytt_android.view.MyFragment
+import com.wonchihyeon.ytt_android.view.OrderListFragment
 
 class MainActivity : AppCompatActivity() {
     private val binding: ActivityMainBinding by lazy {
