+++ FE/app/src/main/java/com/wonchihyeon/ytt_android/data/database/AppDatabase.kt
@@ -1,0 +1,4 @@
+package com.wonchihyeon.ytt_android.data.database
+
+class AppDatabase {
+}
