+++ FE/app/src/main/java/com/wonchihyeon/ytt_android/data/network/ApiService.kt
@@ -1,0 +1,4 @@
+package com.wonchihyeon.ytt_android.data.network
+
+class ApiService {
+}
