+++ FE/app/src/main/java/com/wonchihyeon/ytt_android/view/OrderListFragment.kt
@@ -1,4 +1,4 @@
-package com.wonchihyeon.ytt_android.fragments
+package com.wonchihyeon.ytt_android.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,17 +6,16 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import androidx.navigation.findNavController
 import com.wonchihyeon.ytt_android.R
-import com.wonchihyeon.ytt_android.databinding.FragmentMyBinding
+import com.wonchihyeon.ytt_android.databinding.FragmentOrderBinding
 
-// 사용자 페이지
-class MyFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private lateinit var binding: FragmentMyBinding
 
+// 주문내역
+class OrderListFragment : Fragment() {
+    private lateinit var binding: FragmentOrderBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
     }
 
     override fun onCreateView(
@@ -24,9 +23,11 @@ class MyFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_my, container, false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_order, container, false)
+
 
 
         return binding.root
     }
+
 }
