+++ FE/app/src/main/java/com/wonchihyeon/ytt_android/viewmodel/VendingMachineViewModel.kt
@@ -1,0 +1,4 @@
+package com.wonchihyeon.ytt_android.viewmodel
+
+class VendingMachineViewModel {
+}
