+++ FE/app/src/main/res/layout/activity_order.xml
@@ -4,7 +4,7 @@ xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
 android:orientation="vertical"
-tools:context=".order.OrderDetailsActivity">
+tools:context=".view.order.OrderDetailsActivity">
