+++ FE/app/src/main/res/layout/activity_order_failure.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".order.OrderFailureActivity">
+    tools:context=".view.order.OrderFailureActivity">
 
     <!-- Back button -->
     <ImageButton
